requirements.txt
transformers>=4.30.0
pandas
tqdm
scikit-learn
pillow
opencv-python
numpy
ftfy
regex
clip